JWT几何绘制 判断围栏

+++ dps-atomic/src/main/scala/dps/atomic/util/Point.scala
@@ -1,6 +1,6 @@
 package dps.atomic.util
 
 class Point {
-  var lat:Float = _
-  var lng:Float = _
+  var lat:Double = _
+  var lng:Double = _
 }
